api/python/requirements.txt
flask
python-dotenv
newspaper3k
nltk
